requirements.txt
numpy==1.12.0
pandas==0.19.2
scikit-learn==0.18.1
scipy==0.18.1
PyYAML==3.12
pytest==3.0.5
numba==0.31.0
hypothesis==3.6.1
